+++ src/addons/meta_human_dna/release_notes.md
@@ -3,7 +3,8 @@
 
 ## Patch Changes
 * Fixed disappearing component sub-panels under RigLogic panel [#167](https://github.com/poly-hammer/meta-human-dna-addon/issues/167)
-[#186](https://github.com/poly-hammer/meta-human-dna-addon/issues/186)
+* Fixed fps scale when importing face board animation [#186](https://github.com/poly-hammer/meta-human-dna-addon/issues/186)
+* Added custom operator `File > Import > MetaHuman Link/Append` that allows users to link and append metahuman rigs from other .blend files [#169](https://github.com/poly-hammer/meta-human-dna-addon/issues/169)
 
 
 > [!WARNING]